+++ core/src/serialize/fromText.ts
@@ -1,12 +1,10 @@
 import { ListState } from '../types'
-import valueFromText from './valueFromText'
 import raw from './fromRaw'
 
 const textToFlat = (text: string): ListState => {
   return raw({
     text: `[${text.replace(/\n/gi, '][')}]`,
     ranges: [],
-    entityMap: {}
   })
 
 }