+++ site/src/examples/Changes/index.tsx
@@ -79,7 +79,6 @@ const App = () => {
   const [editorState, setEditorState] = useState(() => EditorState.fromJSON({
     text,
     ranges: [],
-    entityMap: {}
   }))
 
   const onChangeChange = (event: any) => {