+++ site/src/examples/MarkdownTest/index.tsx
@@ -67,7 +67,6 @@ const App = () => {
   const [editorState, setEditorState] = useState(() => EditorState.fromJSON({
     text,
     ranges: [],
-    entityMap: {}
   }))
 
   return (