+++ site/src/examples/TypeStyler/index.tsx
@@ -13,7 +13,6 @@ const App = () => {
   const [editorState, setEditorState] = useState(() => EditorState.fromJSON({
     text,
     ranges: [],
-    entityMap: {}
   }))
 
   const [positions, setPosition] = useState(emptyPositions);